Update Celcius.cs
switch from double to float

+++ TemperatureConverter/Celcius.cs
@@ -10,16 +10,16 @@ namespace TemperatureConverter
     {
         Logger logger = new Logger();
 
-        public double CelciusToFahrenheit(double celcius)
+        public float CelciusToFahrenheit(float celcius)
         {
-            double result = (celcius * 1.8) + 32;
+            float result = (celcius * 1.8f) + 32;
             logger.Log($"{celcius} in Celcius are {result} in Fahrenheit");
             return result;
         }
 
-        public double FahrenheitToCelcius(double fahrenheit)
+        public float FahrenheitToCelcius(float fahrenheit)
         {
-            double result = (fahrenheit - 32) / 1.8;
+            float result = (fahrenheit - 32) / 1.8f;
             logger.Log($"{fahrenheit} in Fahrenheit are {result} in Celcius");
             return result;
         }
